CLN: Total refactor, remove original schema, replace with v2
