+++ interface.go
@@ -3,24 +3,12 @@ package covalent
 import (
 	"github.com/ElrondNetwork/covalent-indexer-go/hyperBlock"
 	"github.com/ElrondNetwork/covalent-indexer-go/schema"
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
-	"github.com/ElrondNetwork/elrond-go-core/data/indexer"
-	vmcommon "github.com/ElrondNetwork/elrond-vm-common"
 	"github.com/elodina/go-avro"
 )
 
-type DataHandler interface {
-	ProcessData(args *indexer.ArgsSaveBlockData) (*schema.BlockResult, error)
-}
-
 // HyperBlockProcessor shall handle hyper block processing into avro schema blocks
 type HyperBlockProcessor interface {
-	Process(hyperBlock *hyperBlock.HyperBlock) (*schemaV2.HyperBlock, error)
-}
-
-type AccountsAdapter interface {
-	LoadAccount(address []byte) (vmcommon.AccountHandler, error)
-	IsInterfaceNil() bool
+	Process(hyperBlock *hyperBlock.HyperBlock) (*schema.HyperBlock, error)
 }
 
 // AvroMarshaller defines what an avro marshaller should do
