+++ process/logs/logs_processor.go
@@ -1,7 +1,7 @@
 package logs
 
 import (
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
+	"github.com/ElrondNetwork/covalent-indexer-go/schema"
 	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
 )
 
@@ -14,19 +14,19 @@ func NewLogsProcessor() *logsProcessor {
 }
 
 // ProcessLog converts logs data to a specific structure defined by avro schema
-func (lp *logsProcessor) ProcessLog(log *transaction.ApiLogs) *schemaV2.Log {
+func (lp *logsProcessor) ProcessLog(log *transaction.ApiLogs) *schema.Log {
 	if log == nil {
-		return schemaV2.NewLog()
+		return schema.NewLog()
 	}
 
-	return &schemaV2.Log{
+	return &schema.Log{
 		Address: []byte(log.Address),
 		Events:  lp.processEvents(log.Events),
 	}
 }
 
-func (lp *logsProcessor) processEvents(events []*transaction.Events) []*schemaV2.Event {
-	allEvents := make([]*schemaV2.Event, 0, len(events))
+func (lp *logsProcessor) processEvents(events []*transaction.Events) []*schema.Event {
+	allEvents := make([]*schema.Event, 0, len(events))
 
 	for _, currEvent := range events {
 		processedEvent := lp.processEvent(currEvent)
@@ -39,12 +39,12 @@ func (lp *logsProcessor) processEvents(events []*transaction.Events) []*schemaV2
 	return allEvents
 }
 
-func (lp *logsProcessor) processEvent(event *transaction.Events) *schemaV2.Event {
+func (lp *logsProcessor) processEvent(event *transaction.Events) *schema.Event {
 	if event == nil {
 		return nil
 	}
 
-	return &schemaV2.Event{
+	return &schema.Event{
 		Address:    []byte(event.Address),
 		Identifier: []byte(event.Identifier),
 		Topics:     event.Topics,
