+++ testscommon/mock/hyper_block_processor_stub.go
@@ -2,16 +2,16 @@ package mock
 
 import (
 	"github.com/ElrondNetwork/covalent-indexer-go/hyperBlock"
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
+	"github.com/ElrondNetwork/covalent-indexer-go/schema"
 )
 
 // HyperBlockProcessorStub -
 type HyperBlockProcessorStub struct {
-	ProcessCalled func(hyperBlock *hyperBlock.HyperBlock) (*schemaV2.HyperBlock, error)
+	ProcessCalled func(hyperBlock *hyperBlock.HyperBlock) (*schema.HyperBlock, error)
 }
 
 // Process -
-func (hbp *HyperBlockProcessorStub) Process(hyperBlock *hyperBlock.HyperBlock) (*schemaV2.HyperBlock, error) {
+func (hbp *HyperBlockProcessorStub) Process(hyperBlock *hyperBlock.HyperBlock) (*schema.HyperBlock, error) {
 	if hbp.ProcessCalled != nil {
 		return hbp.ProcessCalled(hyperBlock)
 	}
