+++ testscommon/mock/log_handler_stub.go
@@ -1,17 +1,17 @@
 package mock
 
 import (
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
+	"github.com/ElrondNetwork/covalent-indexer-go/schema"
 	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
 )
 
 // LogHandlerStub -
 type LogHandlerStub struct {
-	ProcessLogCalled func(log *transaction.ApiLogs) *schemaV2.Log
+	ProcessLogCalled func(log *transaction.ApiLogs) *schema.Log
 }
 
 // ProcessLog -
-func (lhs *LogHandlerStub) ProcessLog(log *transaction.ApiLogs) *schemaV2.Log {
+func (lhs *LogHandlerStub) ProcessLog(log *transaction.ApiLogs) *schema.Log {
 	if lhs.ProcessLogCalled != nil {
 		return lhs.ProcessLogCalled(log)
 	}
