+++ testscommon/mock/receipt_handler_stub.go
@@ -1,17 +1,17 @@
 package mock
 
 import (
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
+	"github.com/ElrondNetwork/covalent-indexer-go/schema"
 	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
 )
 
 // ReceiptHandlerStub -
 type ReceiptHandlerStub struct {
-	ProcessReceiptCalled func(apiReceipt *transaction.ApiReceipt) (*schemaV2.Receipt, error)
+	ProcessReceiptCalled func(apiReceipt *transaction.ApiReceipt) (*schema.Receipt, error)
 }
 
 // ProcessReceipt -
-func (rhs *ReceiptHandlerStub) ProcessReceipt(apiReceipt *transaction.ApiReceipt) (*schemaV2.Receipt, error) {
+func (rhs *ReceiptHandlerStub) ProcessReceipt(apiReceipt *transaction.ApiReceipt) (*schema.Receipt, error) {
 	if rhs.ProcessReceiptCalled != nil {
 		return rhs.ProcessReceiptCalled(apiReceipt)
 	}
