+++ testscommon/processMocks/altered_accounts_handler_stub.go
@@ -2,16 +2,16 @@ package processMocks
 
 import (
 	"github.com/ElrondNetwork/covalent-indexer-go/alteredAccount"
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
+	"github.com/ElrondNetwork/covalent-indexer-go/schema"
 )
 
 // AlteredAccountsHandlerStub -
 type AlteredAccountsHandlerStub struct {
-	ProcessAccountsCalled func(apiAlteredAccounts []*alteredAccount.AlteredAccount) ([]*schemaV2.AccountBalanceUpdate, error)
+	ProcessAccountsCalled func(apiAlteredAccounts []*alteredAccount.AlteredAccount) ([]*schema.AccountBalanceUpdate, error)
 }
 
 // ProcessAccounts -
-func (ahs *AlteredAccountsHandlerStub) ProcessAccounts(apiAlteredAccounts []*alteredAccount.AlteredAccount) ([]*schemaV2.AccountBalanceUpdate, error) {
+func (ahs *AlteredAccountsHandlerStub) ProcessAccounts(apiAlteredAccounts []*alteredAccount.AlteredAccount) ([]*schema.AccountBalanceUpdate, error) {
 	if ahs.ProcessAccountsCalled != nil {
 		return ahs.ProcessAccountsCalled(apiAlteredAccounts)
 	}
