+++ testscommon/processMocks/transaction_handler_stub.go
@@ -1,17 +1,17 @@
 package processMocks
 
 import (
-	"github.com/ElrondNetwork/covalent-indexer-go/schemaV2"
+	"github.com/ElrondNetwork/covalent-indexer-go/schema"
 	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
 )
 
 // TransactionHandlerStub -
 type TransactionHandlerStub struct {
-	ProcessTransactionsCalled func(apiTransactions []*transaction.ApiTransactionResult) ([]*schemaV2.Transaction, error)
+	ProcessTransactionsCalled func(apiTransactions []*transaction.ApiTransactionResult) ([]*schema.Transaction, error)
 }
 
 // ProcessTransactions -
-func (ths *TransactionHandlerStub) ProcessTransactions(apiTransactions []*transaction.ApiTransactionResult) ([]*schemaV2.Transaction, error) {
+func (ths *TransactionHandlerStub) ProcessTransactions(apiTransactions []*transaction.ApiTransactionResult) ([]*schema.Transaction, error) {
 	if ths.ProcessTransactionsCalled != nil {
 		return ths.ProcessTransactionsCalled(apiTransactions)
 	}
